requirements.txt
requests
pandas
pyodbc
azure-eventhub
